requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
anthropic==0.28.0
gunicorn==22.0.0
